feat: add make_channel_max_diff_mask, TID image test tool, and stricter impulse detection - added make_channel_max_diff_mask for per-channel max-difference visualization - added a simple TID testing program that saves original images, cleaned images, and difference masks - improved detect_impulses_row_to_mask with tighter impulse filtering logic

+++ include/iqalab/impulse.hpp
@@ -8,7 +8,7 @@ namespace iqa {
 struct ImpulseStats {
     /// Total number of pixels that were classified as impulses
     /// (i.e. pixels where the output differs from the distorted input).
-    long count = 0;
+    size_t count = 0;
 };
 
 std::size_t count_impulses(const cv::Mat& impulseMask);
